requirements.txt
openai==1.30.1
httpx==0.27.0
feedparser==6.0.11
python-dotenv==1.0.1
fpdf2==2.7.8
pyaes==1.6.1
pyasn1==0.6.1
rsa==4.9.1
Telethon==1.40.0
requests==2.31.0
flask==3.0.3
flask-cors==4.0.1
gunicorn==23.0.0
pywebpush==1.14.0
unidecode==1.3.7
pycountry==22.3.5
numpy==1.24.4
langdetect==1.0.9
concurrent-log-handler==0.9.24
sib-api-v3-sdk==7.6.0
psycopg2-binary==2.9.10
schedule==1.2.2
marshmallow>=3.0
flask-limiter>=3.5
PyJWT>=2.8